+++ native/c/server/schemas/requests.hpp
@@ -203,6 +203,11 @@ ZJSON_SCHEMA(ListDsMembersRequest,
     FIELD_OPTIONAL(pattern, STRING)
 );
 
+struct ResolveDsAliasRequest {};
+ZJSON_SCHEMA(ResolveDsAliasRequest,
+    FIELD_REQUIRED(dsname, STRING)
+);
+
 struct ReadDatasetRequest {};
 ZJSON_SCHEMA(ReadDatasetRequest,
     FIELD_OPTIONAL(stream, ANY),